Trying to get some nice pretty plots for my PC paper

+++ ctraj/scripts/plotting/quick_plot.sh
@@ -1,87 +1,114 @@
 #!/bin/bash
+#set -x
 set -e
 
-#defaults:
+#default values for GMT projection:
 RANGE="0/360/0/90"
 RANGE2="0/360/-90/90"
 PROJ="S0/90/18"
-ORIENT=-ZBLx
 
-NLON=144
-NLAT=73
+#number of longitude and latitude grids:
+NLON=360
+NLAT=181
 
+#dimension of frame:
+XLEN=360
+YLEN=180
+
+#add outlines of landforms:
+VTYPE=0
+
+#vertical grid:
 Z0=-1
 ZF=1
 NZ=21
 
-while getopts 'qc:g:z:I:J:F:Har:x:y:Z:R:2:' ARG; do
+#line thickness for line plots:
+THICK=2
+
+#command for contour plotting:
+CONTOUR=grdimage
+
+#plot continents using lines?
+LFLAG=1
+CUSTOMR=0
+
+while getopts 'c:F:g:I:J:R:T:V:x:X:y:Y:z:HLq' ARG; do
   case $ARG in
-    Z) ORIENT=-Z$OPTARG
-      ;;
-    z) ZOPTS="$ZOPTS -z $OPTARG"
-       NZ=$OPTARG
-       WCFLAG=1
-      ;;
-    I) ZOPTS="$ZOPTS -I $OPTARG"
-       Z0=$OPTARG
-       WCFLAG=1
-      ;;
-    F) ZOPTS="$ZOPTS -F $OPTARG"
-       ZF=$OPTARG
-       WCFLAG=1
-      ;;
     c) PALETTE=$OPTARG
        DDSWTC=1;		
        #dont delete the contour palette before exiting
       ;;
+    F) ZF=$OPTARG
+       WCFLAG=1			#auto-generate colour palette
+      ;;
     g) ZOPTS="$ZOPTS -g"
        WCFLAG=1
       ;;
-    a) AOPT="-a";
+    I) Z0=$OPTARG
+       WCFLAG=1
       ;;
-    r) EXT_OPT="$EXT_OPT -r $OPTARG"
+    J) PROJ=$OPTARG
       ;;
-    x) NLON=$OPTARG
-      ;;
-    y) NLAT=$OPTARG
-      ;;
-    e) CFLAG=1
+    L) CONTOUR=grdcontour
+       LFLAG=0
       ;;
     q) QFLAG=1
       ;;
     R) RANGE=$OPTARG
+       CUSTOMR=1
       ;;
-    2) RANGE2=$OPTARG
+    T) THICK=$OPTARG
       ;;
-    J) PROJ=$OPTARG
+    V) VTYPE=$OPTARG
       ;;
-    H) echo "Usage: quick_plot [-a] [-q] [-r slen] "
+    x) NLON=$OPTARG
+      ;;
+    X) XLEN=$OPTARG
+       CUSTOMRANGE=1
+      ;;
+    y) NLAT=$OPTARG
+      ;;
+    Y) YLEN=$OPTARG
+       CUSTOMRANGE=1
+      ;;
+    z) NZ=$OPTARG
+       WCFLAG=1
+      ;;
+    H) echo "Usage: quick_plot [-q] [-L] [-V 1]"
        echo "            [-z nz] [-I bottom] [-F top] [-g]"
        echo "            [-x nlon] [-y nlat] [outfile]"
        echo "options:"
        echo "  -J projection"
        echo "  -R range"
+       echo "  -L plot lines instead of solid contours"
        echo "  -H help"
        exit 0
        ;;
   esac
 done
 
-ZOPTS="-z $NZ -I $Z0 -F $ZF"
-
 shift $(($OPTIND - 1))
 
-INFILE=/dev/stdin
-#INFILE=dum.in
-#INDEX=$2
-#OUTFILE=dum.ps
+OUTFILE=$1
+
+ZOPTS="$ZOPTS -z $NZ -I $Z0 -F $ZF"
+
+# for non-global and non-projection plots:
+# (this will need to be sorted...)
+if test $CUSTOMRANGE; then
+  RANGE2=0/$XLEN/0/$YLEN
+  if [[ $VTYPE -ne 0 && $CUSTOMR -eq 0 ]]; then
+    RANGE=$RANGE2
+  fi
+fi
 
 NOW=$(date +"%Y.%m.%d_%H-%M-%S")
-BASE=dum.$NOW
+BASE=${OUTFILE%.*}.$NOW
 #BASE=${INFILE%.*}
 #BASE=${BASE##*/}
-OUTFILE=${1:-$BASE.ps}
 
+#colour palette stuff:
 if test -z $DDSWTC; then
   WCFLAG=1;
 fi
@@ -96,28 +123,63 @@ fi
 
 ZGRIDFILE=$BASE.zgrid;
 
-echo "psbasemap -R${RANGE} -J${PROJ} -K -Bg30 > ${OUTFILE}"
-psbasemap -R${RANGE} -J${PROJ} -K -Bg30 > ${OUTFILE}
+if test -z $OUTFILE; then OUTFILE="$BASE.$INDEX.tmp.ps"; fi
 
-  DLON=$(echo "scale=2; 360./$NLON" | bc)
-  DLAT=$(echo "scale=2; 180./($NLAT-1)" | bc)
-  echo $NLON $NLAT $DLON $DLAT
+echo $QFLAG
+echo "psbasemap -R${RANGE} -J${PROJ} -Bg30 -K > ${OUTFILE}"
+psbasemap -R${RANGE} -J${PROJ} -Bg30 -K > ${OUTFILE}
 
-  GRDFILE=$BASE.grd
+echo $QFLAG
 
-  echo "gen_zgrid $ZOPTS > $ZGRIDFILE"
-  gen_zgrid $ZOPTS > $ZGRIDFILE
-  echo "makecpt -T$ZGRIDFILE > $PALETTE"
-  makecpt -T$ZGRIDFILE > $PALETTE
+#add coastlines:
+if test -z $QFLAG; then
+  LFLAG=0
+fi
 
-  echo "xyz2grd -R$RANGE2 -I$DLON/$DLAT $ORIENT -G$GRDFILE < $INFILE"
-  xyz2grd -R$RANGE2 -I$DLON/$DLAT $ORIENT -G$GRDFILE < $INFILE
-  echo "grdimage $GRDFILE -R$RANGE -J$PROJ -C$PALETTE -O -K >> $OUTFILE"
-  grdimage $GRDFILE -R$RANGE -J$PROJ -C$PALETTE -O -K >> $OUTFILE
+if [[ $VTYPE -eq 0 && $LFLAG -eq 0 ]]
+then
+  echo "pscoast -R${RANGE} -J${PROJ} -Dl -K -O >> ${OUTFILE}"
+  pscoast -R${RANGE} -J${PROJ} -G220 -Dl -K -O >> ${OUTFILE}
+fi
 
-echo "pscoast -R${RANGE} -J${PROJ} -Dl -W -O >> ${OUTFILE}"
-pscoast -R${RANGE} -J${PROJ} -Dl -W -O >> ${OUTFILE}
+echo "psbasemap -R${RANGE} -J${PROJ} -K -O -Bg30 >> ${OUTFILE}"
+psbasemap -R${RANGE} -J${PROJ} -K -O -Bg30 >> ${OUTFILE}
 
+if test $QFLAG; then
+  #plot field:
+  DLON=$(date_calc "($XLEN|$NLON)")
+  DLAT=$(date_calc "($YLEN|(${NLAT}_1))")
+
+  GRDFILE=$BASE.$INDEX.grd
+
+  ORIENT=-ZBLx
+
+  #auto-generate colour palette/vertical levels:
+  if test $WCFLAG; then
+    echo "gen_zgrid $ZOPTS > $ZGRIDFILE"
+    gen_zgrid $ZOPTS > $ZGRIDFILE
+    echo "makecpt -T$ZGRIDFILE > $PALETTE"
+    makecpt -T$ZGRIDFILE > $PALETTE
+  fi
+
+  echo "xyz2grd -R$RANGE2 -I$DLON/$DLAT $ORIENT -G$GRDFILE"
+  xyz2grd -R$RANGE2 -I$DLON/$DLAT $ORIENT -G$GRDFILE
+  echo "$CONTOUR $GRDFILE -R$RANGE -J$PROJ -C$PALETTE -O -K >> $OUTFILE;"
+  $CONTOUR $GRDFILE -R$RANGE -J$PROJ -C$PALETTE -O -K >> $OUTFILE;
+else
+  #plot single contour:
+  echo "psxy -R${RANGE} -J${PROJ} -W$THICK,black -O -K >> ${OUTFILE};"
+  psxy -R${RANGE} -J${PROJ} -W$THICK,black -O -K >> ${OUTFILE};
+fi
+
+#add coastlines:
+if [[ $VTYPE -eq 0 && $LFLAG -eq 1 ]]
+then
+  echo "pscoast -R${RANGE} -J${PROJ} -Dl -W -O >> ${OUTFILE}"
+  pscoast -R${RANGE} -J${PROJ} -Dl -W -O >> ${OUTFILE}
+fi
+
+#clean up:
 rm -f $ZGRIDFILE
 
 rm -f $GRDFILE
@@ -128,6 +190,8 @@ if test -z $1; then
 fi
 
 if test -z $DDSWTC; then 
-  rm $PALETTE;
+  if test $QFLAG; then
+	rm $PALETTE;
+  fi;
 fi
